+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-json
 
+## 5.1.6
+
+### Patch Changes
+
+- Updated dependencies [c38faebce]
+  - @inlang/sdk@0.24.0
+
+## 5.1.5
+
+### Patch Changes
+
+- Updated dependencies [b920761e6]
+  - @inlang/sdk@0.23.0
+
 ## 5.1.4
 
 ### Patch Changes